Move ViewComponent:Collection to a dedicated file
